% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleanup.R
\name{add_ref_protein_sequences}
\alias{add_ref_protein_sequences}
\title{Add reference protein sequences}
\usage{
add_ref_protein_sequences(data, path)
}
\arguments{
\item{data}{a data table or tibble with a "protein" column of accession numbers}

\item{path}{path to a TSV containing accession number/sequence pairs}
}
\description{
Add column "protein_sequence" to a tibble by dplyr::left_join() on "protein"
(containing accession numbers)
}
\details{
The source data can be generated from a ReferenceGeneCatalog.txt file using
download_reference_gene_catalog_proteins()
}
