% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{categorize_integer_groups}
\alias{categorize_integer_groups}
\title{Assign column values to groups}
\usage{
categorize_integer_groups(
  data,
  source_col,
  label_col,
  label_sep = " - ",
  groups,
  split = NULL
)
}
\arguments{
\item{source_col}{integer column to group}

\item{label_col}{(new) column for group names}

\item{label_sep}{seperator between upper and lower bondary in lable}

\item{groups}{number integer of groups to define (fed into
determine_nearly_equal_integer_groups_lt() to determine boundaries)}

\item{split}{integer vector of boundaries, which is generally the output of
determine_nearly_equal_integer_groups_lt()}

\item{tibble}{A data frame or tibble}
}
\description{
Companion function to determine_nearly_equal_integer_groups_lt()
}
\details{
Either groups or split should be provided. Split takes precedence.
}
