% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{check_complete_pdg}
\alias{check_complete_pdg}
\title{Verify a PDG accession.version release is complete}
\usage{
check_complete_pdg(
  url = "ftp://ftp.ncbi.nlm.nih.gov/pathogen/Results/",
  organism,
  version
)
}
\arguments{
\item{url}{Path to the NCBI Pathogen Detection FTP Results directory}

\item{organism}{FTP directory name of an organism group included in the NCBI
Pathogen Detection Project}

\item{version}{A PDG accession.version corresponding to \code{organism}}
}
\description{
Verifies that the specified PDG accession.version release is complete (i.e.
includes AMR Metadata, SNP Clusters, and SNP Trees). Note that SNP Trees are
part of the definition of a 'complete' release, but are not downloaded or
analyzed by \code{pdallele}
}
