% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{check_complete_pdg_version}
\alias{check_complete_pdg_version}
\title{Verify a PDG accession.version is complete (FTP)}
\usage{
check_complete_pdg_version(url, organism, version)
}
\arguments{
\item{url}{path to NCBI Pathogen Detection FTP "Results" directory}

\item{organism}{name of a valid organism/organism group}

\item{version}{a valid PDG accession.version}
}
\description{
Verify a PDG accession.version is complete (FTP)
}
