% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleanup.R
\name{clean_filter_alleles}
\alias{clean_filter_alleles}
\title{Clean and filter allele listings}
\usage{
clean_filter_alleles(data, filter, remove)
}
\arguments{
\item{data}{A dataframe or tibble with an allele column}

\item{filter}{A character vector of gene names to keep. Uses grepl(). Partial
matches are kept. Vector is collapsed with "|" (or operator).}

\item{remove}{A character vector of allele_quality values to remove}
}
\description{
allele_quality is a column referring to alleles that are less than perfect
matches, which include HMM derived alleles, end of contig issues, and point
mutations. See possible values and explanations at:
https://www.ncbi.nlm.nih.gov/pathogens/pathogens_help/#genotype-categories
}
\details{
This function also removes any duplicate allele/biosample combinations, which
could complicate later analysis. Distinct alleles are maintained.

For alleles without a gene listed (i.e. unassigned alleles and lower quality
calls not matching an assigned allele), \code{allele} is coalesced into \code{gene}
}
