% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{count_alleles}
\alias{count_alleles}
\title{Count Alleles by Columns}
\usage{
count_alleles(.data, ..., sort = TRUE, name = "alleles", na_last = TRUE)
}
\arguments{
\item{.data}{A data frame or tibble.}

\item{...}{\if{html}{\out{<data masking>}} columns to group by.}

\item{sort}{If TRUE, will show the largest groups at the top.}

\item{na_last}{Where to sort NA. If TRUE, NA is put last; if FALSE,
NA is put first; if NA, NA is dropped.}
}
\description{
This is a convenience function for count_by_column(name = "alleles",
isolates = FALSE)
}
