% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{count_by_column}
\alias{count_by_column}
\alias{count_alleles}
\alias{count_isolates}
\title{Count by columns}
\usage{
count_by_column(
  data,
  ...,
  isolates = FALSE,
  sort = TRUE,
  name = "n",
  na_last = TRUE
)

count_alleles(data, ..., sort = TRUE, name = "alleles", na_last = TRUE)

count_isolates(data, ..., sort = TRUE, na_last = TRUE)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{...}{\if{html}{\out{<data masking>}} columns to group by.}

\item{isolates}{Should isolates (BioSamples) be counted instead of alleles?}

\item{sort}{If TRUE, will show the largest groups at the top.}

\item{name}{The name of the new column in the output.}

\item{na_last}{Where to sort \code{NA}. If \code{TRUE}, \code{NA} is put last; if \code{FALSE},
\code{NA} is put first; if \code{NA}, \code{NA} is dropped.}
}
\description{
Wraps \code{\link[dplyr:count]{dplyr::count()}} and adds additional options for ordering of \code{NA}
values and counting distinct BioSamples

Two convenience functions are also available:
\itemize{
\item \code{count_alleles()} wraps \code{count_by_column(name = "alleles", isolates = FALSE)}
\item \code{count_isolates()} wraps \code{count_by_column(name = "isolates", isolates = TRUE)}
}
}
