% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{determine_combinations}
\alias{determine_combinations}
\title{Add \code{combo} column of combinations of \code{allele} present in each \code{biosample}
and filter out biosamples with only a single allele}
\usage{
determine_combinations(data, filter_terms = NULL)
}
\arguments{
\item{data}{a data frame or tibble}

\item{filter_terms}{character vector of terms to filter alleles before
determining groups}
}
\description{
Add \code{combo} column of combinations of \code{allele} present in each \code{biosample}
and filter out biosamples with only a single allele
}
