% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{determine_exclusive_alleles}
\alias{determine_exclusive_alleles}
\title{Filter alleles exclusive to only a single combination of columns}
\usage{
determine_exclusive_alleles(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{\if{html}{\out{<data-masking>}} cols to find alleles unique to different values}
}
\description{
Filter alleles exclusive to only a single combination of columns
}
