% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{determine_nearly_equal_integer_groups_lt}
\alias{determine_nearly_equal_integer_groups_lt}
\title{Define boundaries to split integer vector into groups of roughly equal size}
\usage{
determine_nearly_equal_integer_groups_lt(values, groups = 4)
}
\arguments{
\item{values}{integer vector of values to group}

\item{groups}{integer of groups to define}
}
\description{
Used to define boundaries to split year into periods
}
\details{
Companion function to categorize_integer_groups()
}
