% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{determine_shared_alleles}
\alias{determine_shared_alleles}
\title{Filter alleles shared between multiple combinations of columns}
\usage{
determine_shared_alleles(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{\if{html}{\out{<data-masking>}} cols to find alleles shared among different values}
}
\description{
Filter alleles shared between multiple combinations of columns
}
