% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut_plot.R
\name{donut_plot}
\alias{donut_plot}
\title{Generate a donut plot}
\usage{
donut_plot(
  data,
  fill,
  counts,
  count_label = "Alleles",
  legend_title = "**Allele**",
  labels = TRUE,
  total = TRUE,
  palette = NULL
)
}
\arguments{
\item{data}{A data table or tibble}

\item{fill}{Column containing levels/labels (e.g. alleles names) (rich text)}

\item{counts}{Column containing counts for the levels}

\item{count_label}{Label for counts (e.g. Alleles) (rich text)}

\item{legend_title}{Label for the legend (rich text)}

\item{labels}{Should callouts with count & percentage be shown for each arc?}

\item{total}{Should total of \code{count} be displayed inside the donut?}

\item{palette}{A \code{\link[ggthemes:tableau_color_pal]{ggthemes::tableau_color_pal()}} color palette name}
}
\description{
Generates a donut plot
}
\seealso{
\code{\link[=donut_plot_panel]{donut_plot_panel()}} to generate a faceted panel containing multiple
donut plots
}
