% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut_plot.R
\name{donut_plot_panel}
\alias{donut_plot_panel}
\title{Generates a faceted panel of donut plots}
\usage{
donut_plot_panel(
  data,
  fill,
  counts,
  facet,
  count_label = "Alleles",
  legend_title = "**Allele**",
  total = TRUE,
  palette = NULL,
  nrow = NULL,
  ncol = NULL
)
}
\arguments{
\item{data}{A data table or tibble}

\item{fill}{Column containing levels/labels (e.g. alleles names) (rich text)}

\item{counts}{Column containing counts for the levels}

\item{facet}{column to facet plots by (e.g. region)}

\item{count_label}{Label for counts (e.g. Alleles) (rich text)}

\item{legend_title}{Label for the legend (rich text)}

\item{total}{Should total of \code{count} be displayed inside the donut?}

\item{palette}{A \code{\link[ggthemes:tableau_color_pal]{ggthemes::tableau_color_pal()}} color palette name}

\item{nrow}{number of facet rows}

\item{ncol}{number of facet columns}
}
\description{
Generates a faceted panel of donut plots
}
\seealso{
\code{\link[=donut_plot]{donut_plot()}} to generate a single donut plot
}
