% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.resolve_na_last}
\alias{.resolve_na_last}
\title{Process na_last arguments}
\usage{
.resolve_na_last(.data, ..., na_last)
}
\arguments{
\item{.data}{a dataframe or tibble}

\item{...}{\if{html}{\out{<data-masking>}} columns to reorder NAs}

\item{na_last}{argument to be processed (pass directly from parent function)}
}
\description{
Used internally to avoid replicating na_last logic across multiple functions
}
