% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{download_mlst_profile}
\alias{download_mlst_profile}
\title{Download and save MLST schema}
\usage{
download_mlst_profile(path, url)
}
\arguments{
\item{path}{path to save schema to}

\item{url}{path to download schema from}
}
\description{
Download and save MLST schema
}
