% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{download_protein_sequences}
\alias{download_protein_sequences}
\title{Download protein sequences from NCBI}
\usage{
download_protein_sequences(id, n = 100)
}
\arguments{
\item{id}{character vector of NCBI protein accession numbers}

\item{n}{how many sequences should be downloaded in a single request?}
}
\description{
Easily download a set of protein sequences by accession number using the
Eutilities web interface via the REntrez package
}
