% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_reference_gene_catalog}
\alias{download_reference_gene_catalog}
\title{Download Reference Gene Catalog}
\usage{
download_reference_gene_catalog(
  url = "ftp://ftp.ncbi.nlm.nih.gov/pathogen/Antimicrobial_resistance/Data/",
  version,
  path
)
}
\arguments{
\item{url}{Path to the NCBI Pathogen Detection FTP Reference Gene Catalog
directory}

\item{version}{A Reference Gene Catalog version number}

\item{path}{Destination directory}
}
\description{
Downloads a specified \code{version} of the Reference Gene Catalog from the NCBI
Pathogen Detection FTP server.

The downloaded file is saved as 'ReferenceGeneCatalog.tsv' in \code{path}.
}
