% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{download_reference_gene_catalog_proteins}
\alias{download_reference_gene_catalog_proteins}
\title{Download (and save) all Reference Gene Catalog protein sequences from NCBI}
\usage{
download_reference_gene_catalog_proteins(
  path = refgene_path,
  file = "reference_proteins.tsv"
)
}
\arguments{
\item{path}{path to an NCBI ReferenceGeneCatalog.txt file}

\item{file}{filename to save the accession/sequence pairs}
}
\description{
Easily download a set of protein sequences by accession number using the
Eutilities web interface via the REntrez package
}
