% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_refgene_catalog}
\alias{download_refgene_catalog}
\title{List available ReferenceGeneCatalog versions (FTP)}
\usage{
download_refgene_catalog(url, version, path)
}
\arguments{
\item{url}{path to NCBI Pathogen Detection FTP Reference Gene Catalog directory}

\item{version}{a valid Reference Gene Catalog version number}

\item{path}{directory to save downloaded files to}
}
\description{
List available ReferenceGeneCatalog versions (FTP)
}
