% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ensure_directory}
\alias{ensure_directory}
\title{Ensure directory exists}
\usage{
ensure_directory(path)
}
\arguments{
\item{path}{Directory path}
}
\description{
Checks if the directory path provided exists and creates it (with
notification) if it does not.
}
