% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{fastmlst_scheme}
\alias{fastmlst_scheme}
\title{Determine FastMLST species 'codename' R implementation of
best_guess_codename(text) from FastMLST v0.0.15. Used to match the "scheme"
listing of FastMLST CSV to PubMLST schema}
\usage{
fastmlst_scheme(text)
}
\arguments{
\item{text}{species name to parse}
}
\description{
See: https://github.com/EnzoAndree/FastMLST
}
