% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cleanup.R
\name{filter-alleles}
\alias{filter-alleles}
\alias{filter_isolates_browser}
\alias{filter_microbigge}
\title{Filter alleles to prepare data set for analysis}
\usage{
filter_isolates_browser(data, filter, remove, deduplicate = TRUE)

filter_microbigge(
  data,
  coverage = 100L,
  identity = 90L,
  filter,
  remove,
  deduplicate = TRUE
)
}
\arguments{
\item{data}{A dataframe or tibble with an \code{allele} column}

\item{filter}{A character vector of \code{allele} names to keep. Uses regex,
vector is collapsed with "|".}

\item{remove}{A character vector of \code{method} values to remove. Uses regex,
vector is collapsed with "|".}

\item{deduplicate}{Should duplicate \code{allele}/\code{biosample} combinations be
removed?}

\item{coverage}{Minimum percent coverage to keep}

\item{identity}{Minimum percent identity to keep}
}
\value{
\code{data} with rows removed according to the selected parameters
}
\description{
Both functions serve three major roles in preparing data sets
for analysis, all of which are optional:
\itemize{
\item keep rows with \code{allele} matching a value in \code{filter}
\item drop rows with \code{method} matching a value in \code{remove}
\item drop rows with duplicate \code{allele}/\code{biosample} combinations based on a logical value \code{deduplicate}
}

Additionally, \code{filter_microbigge()} will drop rows not meeting minimum
thresholds of \code{coverage} and \code{identity}.
\subsection{Allele calling method The \code{method} column contains codes representing}{

the calling method used for a given allele, some of which correspond to
"less than perfect" matches and may warrant removal depending on the scope
and goals of the analysis. Detailed explanations are available at:
\link{https://www.ncbi.nlm.nih.gov/pathogens/pathogens_help/#genotype-categories}
}

\subsection{Coverage and identity These values are produced by AMRFinderPlus as part}{

of the analysis pipeline:
\itemize{
\item \code{coverage} is the percentage of the reference sequence covered by the alignment between the target and the reference sequence
\item \code{identity} is the percentage of identical amino acids or base pairs in the alignment between the target and reference sequence
}

The default thresholds correspond to those used in Isolates Browser (and
thus \code{amr.metadata.tsv} files on the NCBI FTP server). See more info at:
\link{https://www.ncbi.nlm.nih.gov/pathogens/pathogens_help/#genotype-categories}
}

\subsection{Deduplication}{

Removes duplicate \code{allele}/\code{biosample} combinations, which can be included
for a variety of reasons. Runs \code{dplyr::distinct(biosample, allele, .keep_all = TRUE)}.
}
}
