% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{filter_attribute}
\alias{filter_attribute}
\title{Filter attribute column by string}
\usage{
filter_attribute(data, attribute, string)
}
\arguments{
\item{data}{A data frame or tibble}

\item{attribute}{\if{html}{\out{<data-masking>}} Column to filter}

\item{string}{String to look for}
}
\description{
A convenience function wrapping \code{filter(grepl(string, attribute, ignore.case = TRUE))} which looks cleaner in pipes and follows tidyverse argument order
}
