% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{filter_exclusive_alleles}
\alias{filter_exclusive_alleles}
\alias{filter_shared_alleles}
\title{Filter by allele exclusivity}
\usage{
filter_exclusive_alleles(data, ...)

filter_shared_alleles(data, ...)
}
\arguments{
\item{data}{A data frame or tibble}

\item{...}{\if{html}{\out{<data-masking>}} Columns to determine exclusivity by}
}
\value{
A data frame or tibble with only exclusive
(\code{filter_exclusive_alleles}) or shared (\code{filter_shared_alleles}) alleles
}
\description{
Filter \code{data} to keep alleles based on whether they are exclusive to one
group or shared between multiple groups. Two variants are available:
\itemize{
\item \code{filter_exclusive_alleles()} keeps alleles appearing only in one group of \code{group_by(data, ...)}
\item \code{filter_shared_alleles()} keeps alleles appearing in multiple groups of \code{group_by(data, ...)}
}
}
