% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleanup.R
\name{filter_microbigge}
\alias{filter_microbigge}
\title{Filter imported MBE data to remove less "reliable" protein calls}
\usage{
filter_microbigge(
  .data,
  min_coverage = 100,
  min_identity = 90,
  remove = remove_mbe_allele_types
)
}
\arguments{
\item{.data}{a data tabel or tibble}

\item{coverage}{integer minimum percent coverage to keep}

\item{identity}{integer minimum percent identity to keep}
}
\description{
Also allows for the customization of coverage and identity thresholds. The
default identity of 90\% is chosen to match the Isolates Browser criteria
as stated at:
https://www.ncbi.nlm.nih.gov/pathogens/pathogens_help/#genotype-categories
}
\details{
********** THIS REALLY NEEDS TO BE REVISITED **********
}
