% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleanup.R
\name{filter_pds}
\alias{filter_pds}
\title{Filter one biosample for each pds/genotype combination (while also keeping
non-clustered biosamples)}
\usage{
filter_pds(data, ignore_genotype = FALSE)
}
\arguments{
\item{data}{A dataframe or tibble}

\item{ignore_genotype}{should only a single biosample per cluster be kept,
even when multiple genotypes exist?}
}
\description{
Filters data to keep only a single biosample from each pds/genotype (in this
context, the combination of alleles present in a biosample) combination while
also keeping all biosamples without a pds cluster assigned.
}
\details{
This may allow one to focus more on broader diversity by removing duplicates
of successful clones from the dataset
}
