% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cleanup.R
\name{filter_required_alleles}
\alias{filter_required_alleles}
\title{Filter isolates by presence of genes and blaOXA Families}
\usage{
filter_required_alleles(data, genes, family)
}
\arguments{
\item{data}{A data frame or tibble}

\item{genes}{A character vector of (potentially partial) gene names
to keep. Collapsed with '|' and filtered with \code{\link[=grepl]{grepl()}} on \code{allele}.}

\item{family}{A character vector of (potentially partial) blaOXA family names
to keep. Collapsed with '|' and filtered with \code{\link[=grepl]{grepl()}} on \code{oxa_family}.}
}
\description{
Keep only isolates containing the gene(s) and blaOXA family (or families) listed in the
character vectors \code{genes} and  \code{family}. Filters based on BioSamples.
}
