% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_reference_gene_catalogs}
\alias{get_reference_gene_catalogs}
\title{Download listing of available Reference Gene Catalog releases}
\usage{
get_reference_gene_catalogs(
  url = "ftp://ftp.ncbi.nlm.nih.gov/pathogen/Antimicrobial_resistance/Data/"
)
}
\arguments{
\item{url}{Path to the NCBI Pathogen Detection FTP Reference Gene Catalog
directory}
}
\value{
A tibble with columns \code{version} containing available Reference Gene
Catalog releases and \code{updated} containing the date of the most recent
update.
}
\description{
Downloads listing of Reference Gene Catalog releases currently available on
the NCBI Pathogen Detection FTP server.
}
