% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_refgene_catalogs}
\alias{get_refgene_catalogs}
\title{List available ReferenceGeneCatalog versions (FTP)}
\usage{
get_refgene_catalogs(url = ftp_refgene_base)
}
\arguments{
\item{url}{path to NCBI Pathogen Detection FTP Reference Gene Catalog directory}
}
\description{
List available ReferenceGeneCatalog versions (FTP)
}
