% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_versions}
\alias{get_versions}
\title{List Available accession.versions for an Organism (FTP)}
\usage{
get_versions(url, organism)
}
\arguments{
\item{url}{path to NCBI Pathogen Detection FTP "Results" directory}

\item{organism}{name of a valid organism/organism group}
}
\description{
List Available accession.versions for an Organism (FTP)
}
