% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{heatmap_data}
\alias{heatmap_data}
\title{Prepare data for rank and proprtion heatmap}
\usage{
heatmap_data(
  .data,
  count = allele_markdown,
  group,
  n = 5,
  facet,
  by_group = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{.data}{a dataframe or tibble}

\item{count}{\if{html}{\out{<data-masking>}} column to count (long axis)}

\item{group}{\if{html}{\out{<data-masking>}} column to compare across (short axis)}

\item{n}{number of most frequent \code{count} to include (i.e. top n =)}

\item{facet}{\if{html}{\out{<data-masking>}}column to facet by}

\item{by_group}{should \code{n} be aggregated by group or for the overall dataset?}

\item{na.rm}{should NA values be removed from \code{count}, \code{group}, and \code{facet}?}
}
\description{
Count all instances of \code{count} by \code{group}, filter \code{n} most frequent \code{count}
either overall or \code{by_group}. Optionally prepare to facet by \code{facet} and
remove NA values.
}
\details{
Proportion refers to the proportion of isolates in a given \code{group} associated
with each \code{count}. Ranks are the \code{dense_rank(desc())} of the number of
isolates associated with each \code{count}.

Note: The default setting \code{by_group = TRUE} will keep all \code{count} that appear
in the \code{n} most frequent in any \code{group}, typically leading to more than \code{n}
total values being kept. Conversely, \code{by_group = FALSE} will keep only the
\code{n} most frequent from the entire tibble, ensuring only \code{n} values in the
output.
}
