% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap_plot}
\alias{heatmap_plot}
\title{Generate heat map plot}
\usage{
heatmap_plot(
  data,
  count,
  category,
  facet,
  option = "viridis",
  wide = FALSE,
  rank = "**Rank**",
  prop = "**Proportion**",
  prop_breaks = c(1e-04, 5e-04, 0.001, 0.0025, 0.005, 0.01, 0.025, 0.05, 0.1, 0.2, 0.3,
    0.5, 0.75, 0.9),
  prop_labels = c("0.01\%", "0.05\%", "0.1\%", "0.25\%", "0.5\%", "1\%", "2.5\%", "5\%",
    "10\%", "20\%", "30\%", "50\%", "75\%", "90\%"),
  prop_barheight = 15,
  prop_limits = NULL,
  rank_breaks = c(1, 5, 10, 25, 50, 75, 100, 150, 200),
  rank_limits = NULL,
  rank_range = c(12, 1),
  text_size = 16
)
}
\arguments{
\item{data}{A dataframe or tibble}

\item{count}{\if{html}{\out{<data-masking>}} Column of counts (long axis)}

\item{category}{\if{html}{\out{<data-masking>}} Column to compare across (short axis)}

\item{facet}{Column to facet by.}

\item{option}{Color palette, passed to \code{viridis::scale_color_viridis()}.}

\item{wide}{Should the output be wide or tall? See description.}

\item{rank}{Title for the rank (size) legend}

\item{prop}{Title for the proportion (color) legend}

\item{prop_breaks}{Proportion values to be displayed in the legend. Passed to
\code{ggplot2::continuous_scale(breaks = )}}

\item{prop_labels}{Labes corresponding to \code{prop_breaks}. Passed to
\code{ggplot2::continuous_scale(labels = )}}

\item{prop_barheight}{Height of the bar for the proportion legend Passed to
\code{ggplot2::guide_colorbar(barheight = )}}

\item{prop_limits}{Limits of the proportion color scale. Passed to
\code{ggplot2::continuous_scale(limits = )}}

\item{rank_breaks}{Rank values to be displayed in the legend. Passed to
\code{ggplot2::scale_size(breaks = )}.}

\item{rank_limits}{Limits of the rank size scale. Passed to
\code{ggplot2::scale_size(limits = )}}

\item{rank_range}{Minimum and maximum size of the circle. Passed to
\code{ggplot2::scale_size(range = )}}

\item{text_size}{Text size in plot. Passed to \code{ggplot2::element_text()}.}
}
\value{
a \code{ggplot2} plot object
}
\description{
Generate a heatmap plot from a tibble prepared by \code{\link[=heatmap_data]{heatmap_data()}}.

Rank corresponds to size of each circle and proportion corresponds to the
color of each circle.

Plot orientation:
\itemize{
\item \code{wide = FALSE} places \code{count} on the Y-axis and \code{category} on the X-axis which
will result in "longer" plots when \code{count} has more levels than \code{category}.
Plots will facet vertically.
\item \code{wide = TRUE} swaps the axes and will result in "wider" plots when \code{count}
has more levels than \code{category}. Plots will facet horizontally.
}

Facet directions are fixed with plot orientation to avoid unexpected,
unintended, and undesirable "tiling" of plots
}
\seealso{
\code{\link[=heatmap_data]{heatmap_data()}} to prepare data for plotting as a heatmap
}
