% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleanup.R
\name{import_cluster_list}
\alias{import_cluster_list}
\title{Import and Add NCBI Pathogen Detection SNP Cluster Data}
\usage{
import_cluster_list(.data, path)
}
\arguments{
\item{path}{Path to an NCBI Pathogen Detection
PDG_accession.version.reference_target.cluster_list.tsv file}
}
\description{
Import and Add NCBI Pathogen Detection SNP Cluster Data
}
