% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleanup.R
\name{import_ipg}
\alias{import_ipg}
\title{Import Identical Protein Groups data}
\usage{
import_ipg(data, path)
}
\arguments{
\item{data}{A dataframe or tibble}

\item{path}{path to an NCBI IPG tsv file}
}
\value{
\code{data} with columns \code{ipg}, \code{ipg_accession}, and \code{ipg_name} added
}
\description{
Imports Identical Protein Groups data from \code{path} and adds
columns corresponding to the IPG UID (\code{ipg}), IPG Accession Number
(\code{ipg_accession}), and name of the IPG protein (\code{iph_name}) to \code{data}.
}
