% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleanup.R
\name{import_isolates_browser_metadata}
\alias{import_isolates_browser_metadata}
\title{Import, select, and trim NCBI Pathogen Detection Isolates Browser Metadata}
\usage{
import_isolates_browser_metadata(path)
}
\arguments{
\item{path}{Path to an NCBI Isolates Browser 'amr.metadata.tsv' file}
}
\description{
Imports Isolates Browser Metadata, selects and renames relevant columns, and
cleans up problematic quotation marks
}
