% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleanup.R
\name{import_microbigge}
\alias{import_microbigge}
\title{Import and Select Columns from NCBI Pathogen Detection MicroBIGG-E File}
\usage{
import_microbigge(path)
}
\arguments{
\item{path}{Path to an NCBI MicroBIGG-E csv/tsv file.}
}
\description{
Imports NCBI Pathogen DetectionMicrobial Browser for Identification of
Genetic and Genomic Elements (MicroBIGG-E) data and selects and renames
desired columns for downstream processing steps
}
\details{
As of October 2022, up to 100,000 rows of data can be downloaded via the web
interface at: https://www.ncbi.nlm.nih.gov/pathogens/microbigge/

Larger sets can be downloaded using Google Big Query. See instructions from
NCBI at https://www.ncbi.nlm.nih.gov/pathogens/docs/microbigge_gcp/
}
