% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleanup.R
\name{import_regions}
\alias{import_regions}
\title{Adds Region information based on a file and the location_broad column}
\usage{
import_regions(data, path)
}
\arguments{
\item{data}{A dataframe or tibble}

\item{path}{path to a csv/tsv file containing columns 'country' and 'region'}
}
\description{
Requires a csv/tsv file with a 'country' and 'region' column. Countries are
matched to location_broad using countrycode::countryname to account for names
given in as many languages as possible.
}
\details{
This function will also append any other columns in the file to the dataset
(using the column name from the file). This allows the inclusion of multiple
regions or including a region_markdown or similar column
}
