% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_process.R
\name{parse_gene_names}
\alias{parse_gene_names}
\title{Parse MBE \code{gene} names from \code{allele}}
\usage{
parse_gene_names(data)
}
\arguments{
\item{data}{A data frame or tibble containing an \code{allele} column}
}
\value{
\code{data} with column \code{gene} added
}
\description{
Uses regular expressions to extract gene names from allele names
(e.g., by removing asisgned numbers)
}
