% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleanup.R
\name{parse_genus_species}
\alias{parse_genus_species}
\title{Parse scientific_name to genus and species}
\usage{
parse_genus_species(data)
}
\arguments{
\item{path}{Path to an NCBI Isolates Browser
PDG_accession.version.metadata.tsv file}
}
\description{
Parses scientific_name into genus and species columns and their formatted
equivalents. Unpublished "\if{html}{\out{<Genus>}} sp. <strain_identifier>" (See PMC7408187)
are treated as genus = \if{html}{\out{<Genus>}} and species = sp.
}
\details{
See "How Many Species" in PMC3245000 for info in "sp." strains
Also PMC7408187 (Curation of prokaryotes) and
https://www.ncbi.nlm.nih.gov/biosample/docs/organism/
And https://ena-docs.readthedocs.io/en/latest/faq/taxonomy_requests.html
}
