% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleanup.R
\name{parse_ib_oxa_family}
\alias{parse_ib_oxa_family}
\title{Add oxa_family parsed from name (originally from Reference Gene Catalog file)}
\usage{
parse_ib_oxa_family(.data)
}
\arguments{
\item{.data}{A dataframe or tibble}
}
\description{
Also adds markdown and math formatted oxa_family columns
}
\details{
Returns the original table if no OXA alleles are present
}
