% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_process.R
\name{parse_intrinsic_acquired}
\alias{parse_intrinsic_acquired}
\title{Parse whether \code{gene} and \code{oxa_family} are likely intrinsic or acquired}
\usage{
parse_intrinsic_acquired(data, genes = c(), oxa_families = c())
}
\arguments{
\item{data}{A data frame or tibble containing an \code{allele} column}

\item{genes}{character vector of intrinsic genes}

\item{oxa_families}{character vector of intrinsic OXA families}
}
\value{
\code{data} with logical column \code{gene_type} added
}
\description{
Generates a logical column \code{gene_type} listing whether an allele
is \code{intrinsic} or \code{acquired} based on input lists of \code{genes} and
\code{oxa_families} considered to be intrinsic.
}
