% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleanup.R
\name{parse_mbe_oxa_family}
\alias{parse_mbe_oxa_family}
\title{Determines OXA family from MBE data and adds a 'family' column}
\usage{
parse_mbe_oxa_family(data, by = "allele")
}
\arguments{
\item{data}{A dataframe or tibble}

\item{by}{passed to left_join as a join by command; the column(s) to match in the join}

\item{.refgene}{A dataframe or timble imported from an NCBI ReferenceGeneCatalog file}
}
\description{
Determines OXA family from MBE data and adds a 'family' column
}
