% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{parse_mlst_profiles}
\alias{parse_mlst_profiles}
\title{Determine MLST based on allele calls and profiles}
\usage{
parse_mlst_profiles(.calls, .profiles)
}
\arguments{
\item{.calls}{a datatable or tibble of FastMLST allele calls}

\item{.profiles}{a datatable or tibble of PubMLST MLST profiles}
}
\description{
Determine MLST based on allele calls and profiles
}
