% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleanup.R
\name{parse_year}
\alias{parse_year}
\title{Add year parsed from collection_date}
\usage{
parse_year(.data)
}
\arguments{
\item{.data}{A dataframe or tibble}
}
\description{
Add year parsed from collection_date
}
