% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{pipe_inform}
\alias{pipe_inform}
\title{Pass data along the pipeline and inform}
\usage{
pipe_inform(data, ...)
}
\arguments{
\item{data}{a data frame or tibble}

\item{...}{parameters to pass to inform()}
}
\value{
unaltered \code{data}
}
\description{
Pass data along the pipeline and invoke rlang::inform() with the desired message
}
\seealso{
\code{\link[=pipe_warn]{pipe_warn()}}
}
