% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{pipe_warn}
\alias{pipe_warn}
\title{Pass data along the pipeline and rlang::warn()}
\usage{
pipe_warn(.data, ...)
}
\arguments{
\item{.data}{a dataframe or tibble}

\item{...}{parameters to pass to warn()}
}
\description{
Pass data along the pipeline and rlang::warn()
}
