% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{possible_unique_proteins}
\alias{possible_unique_proteins}
\title{Determine all possible unique proteins}
\usage{
possible_unique_proteins(data)
}
\arguments{
\item{data}{A data frame or tibble with columns \code{allele} and \code{protein}}
}
\value{
A character vector of accession numbers
}
\description{
Uses \code{protein} accession numbers and \code{allele} names to create a
character vector representing the complete subset of \emph{possible} unique
alleles in the datasets.
}
\details{
This function is only for use with MicroBIGG-E datasets, as Isolates
Browser datasets do not contain sufficient data to make the determination
of unique unassinged alleles.

Every distinct assigned allele with a "WP_" prefixed accession number is
deduplicated. This does NOT alter values in the original \code{data} and some
downstream matching may require Identical Protein Groups data.

Every distinct accession number corresponding to an unassigned allele is
included as they correspond to \emph{possible} distinct alleles. In most
datasets, there will likely be overlap with several of these accession
numbers corresponding to the same distinct protein. Downstream use may
require deduplication, which can be achieved using Identical Protein Groups
data.
}
