% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{possible_unique_proteins_mbe}
\alias{possible_unique_proteins_mbe}
\title{Determine all possible unique proteins present in an MicroBIGG-E dataset}
\usage{
possible_unique_proteins_mbe(data)
}
\arguments{
\item{data}{a tibble or dataframe containing MicroBIGG-E data}
}
\description{
Combines a single accession number for each distinct, assigned allele with
all distinct accession numbers for unassigned alleles into a single character
vector representing the maximum number of potential distinct alleles present
in the input dataset
}
