% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{process_mlst}
\alias{process_mlst}
\title{Decode and simplify FastMLST ST calls}
\usage{
process_mlst(mlst)
}
\arguments{
\item{mlst}{tibble or data frame of raw FastMLST csv input}
}
\description{
Decode and simplify FastMLST ST calls
}
