% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pubmlst_listing}
\alias{pubmlst_listing}
\title{Download list of all PubMLST MLST schema}
\usage{
pubmlst_listing(url = "https://pubmlst.org/static/data/dbases.xml")
}
\arguments{
\item{url}{(string) URL of PubMLST "dbases.xml" (this should only be needed
if the site changes unexpectedly)}
}
\description{
Download list of all PubMLST MLST schema
}
