% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{relevel_first_last}
\alias{relevel_first_last}
\title{Relevel to first/last position and sort}
\usage{
relevel_first_last(data, ..., first = NULL, last = NULL, na_last = TRUE)
}
\arguments{
\item{data}{A data frame or tibble}

\item{...}{\if{html}{\out{<data masking>}} columns to reorder}

\item{first}{a character vector of values to move to the first position}

\item{last}{a character vector of values to move to the last position}

\item{na_last}{Where should NA go? TRUE at the end, FALSE at the beginning, NA dropped.}
}
\description{
Relevel character vector values to first or last position(s) in \code{...} columns
}
\details{
Note that both \code{first} and \code{last} will remain in the order provided
}
