% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{relevel_numeric}
\alias{relevel_numeric}
\title{Relevel factors in human readable order and sort columns}
\usage{
relevel_numeric(.data, ...)
}
\arguments{
\item{.data}{A dataframe or tibble}

\item{...}{\if{html}{\out{<data-masking>}} cols to relevel alphanumerically}
}
\description{
Relevel factors in 'human' alphanumeric order(i.e. digits are sorted
numerically rather than as strings) and sort columns
}
