% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{remove_unassigned_bla}
\alias{remove_unassigned_bla}
\title{Remove unassigned bla alleles}
\usage{
remove_unassigned_bla(.data)
}
\arguments{
\item{.data}{A data frame or tibble}
}
\description{
Filter .data to remove unassigned bla alleles
(i.e. those with an allele number)
}
