% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{reparse_mlst}
\alias{reparse_mlst}
\title{Automatically re-parse FastMLST calls using profiles downloaded from PubMLST}
\usage{
reparse_mlst(path, save = TRUE)
}
\arguments{
\item{path}{location of FastMLST csv output file}

\item{save}{should the re-parsed file be saved?}
}
\description{
Download correct schema from PubMLST and use to re-parse allele calls from
FastMLST. Optionally save the re-parsed file for future use.
}
\details{
This is primarily useful to handle situations where multiple alleles present
in a single strain caused failed MLST typing in the original file.
}
