% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{reprocess_mlst}
\alias{reprocess_mlst}
\title{Automatically reprocess FastMLST calls using profiles downloaded from PubMLST}
\usage{
reprocess_mlst(path, save = TRUE)
}
\arguments{
\item{path}{location of FastMLST csv output file}

\item{save}{should the reprocessed file be saved?}
}
\description{
Download schema from PubMLST and use to reprocess allele calls from
FastMLST. Optionally save the reprocessed file for future use.

This is primarily useful to handle situations where multiple alleles present
in a single strain caused failed MLST typing in the original file.
}
