% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{select_ftp_organism}
\alias{select_ftp_organism}
\title{Interactively select an organism from the NCBI Pathogen Detection FTP directory}
\usage{
select_ftp_organism(url)
}
\arguments{
\item{url}{path to NCBI Pathogen Detection FTP "Results" directory}
}
\description{
Interactively select an organism from the NCBI Pathogen Detection FTP directory
}
\seealso{
Other ftp download helpers: 
\code{\link{select_ftp_refgene}()},
\code{\link{select_ftp_version}()}
}
\concept{ftp download helpers}
