% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{select_ftp_refgene}
\alias{select_ftp_refgene}
\title{Interactively select a ReferenceGeneCatalog version from the NCBI Pathogen
Detection FTP directory}
\usage{
select_ftp_refgene(url)
}
\arguments{
\item{url}{path to NCBI Pathogen Detection FTP Reference Gene Catalog directory}
}
\description{
Interactively select a ReferenceGeneCatalog version from the NCBI Pathogen
Detection FTP directory
}
