% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{select_ftp_version}
\alias{select_ftp_version}
\title{Interactively select an accession.version for a given organism from the NCBI
Pathogen Detection FTP directory}
\usage{
select_ftp_version(url = ftp_results_base, organism)
}
\arguments{
\item{url}{path to NCBI Pathogen Detection FTP "Results" directory}

\item{organism}{name of a valid organism/organism group}
}
\description{
Interactively select an accession.version for a given organism from the NCBI
Pathogen Detection FTP directory
}
