% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleanup.R
\name{separate_genotypes}
\alias{separate_genotypes}
\title{Separate genotypes (amr, stress, vir) by row to allele and allele_type}
\usage{
separate_genotypes(data, include = "amr")
}
\arguments{
\item{data}{A dataframe or tibble with genotype columns (amr, stress, vir)}
}
\description{
Uses separate_rows to separate genotype columns (amr, stress, vir) originally
AMR_genotype, stress_genotype, and virulance_genotype in isolates browser.
}
\details{
allele_type -> amr, stress, or vir (virulence)
allele_quality -> codes correcponding to reasons for non-"complete" alleles,
see description of clean_filter_alleles() for details
}
