% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{separate_rows_sequential}
\alias{separate_rows_sequential}
\title{Apply separate_rows sequentially by column}
\usage{
separate_rows_sequential(.data, ..., sep = "[^[:alnum:].]+", convert = FALSE)
}
\arguments{
\item{.data}{a dataframe or tibble}

\item{...}{\if{html}{\out{<tidy-select>}} columns to apply separate_rows to}

\item{sep}{Separator delimiting collapsed values.}

\item{convert}{If TRUE will automatically run type.convert() on the key
column. This is useful if the column types are actually numeric, integer,
or logical.}
}
\description{
Useful for separating and all unique combinations of separated rows when
different columns may have different numbers of delimiters for the same row
(which causes a vector recycling error when applying separate_rows to
multiple columns). This essentially "expands" the data column by column.
}
\details{
This approach is relatively slow, but serves the purpose without needing to
modify separate_rows directly.
}
