% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleanup.R
\name{split_location}
\alias{split_location}
\title{Splits location into location_broad (country) and location_detail (city/area)}
\usage{
split_location(data)
}
\arguments{
\item{data}{A dataframe or tibble}
}
\description{
Splits INSDC/BioSample geo_loc_name style country:detailed_location format
into location_broad and location_detail.
}
