% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{standardize_countries}
\alias{standardize_countries}
\title{Standardize country names}
\usage{
standardize_countries(data, custom_match = NULL)
}
\arguments{
\item{data}{A character vector of country names}

\item{custom_match}{A named character vector correlating original country
names to standardized country names. Passed to \code{custom_match} of
\code{\link[countrycode:countrycode]{countrycode::countrycode()}}.}
}
\value{
A tibble with \code{original} and \code{standardized_country} names
}
\description{
Uses \code{countrycode} package to attempt to fix spelling errors and account for
non-English country names in order to "standardize" spelling of country
names. Allows the use of \code{custom_match} to deal with INSDC names that
\code{countrycode} doesn't handle well.

Note that only clearly distinguishable entities with an ISO 3166-1 Alpha-3
code are included (meaning oceans, seas, some islands without ISO codes, some
islands shared between multiple countries (e.g. Borneo), and some disputed
territories are dropped).
}
