% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\name{theme_simple}
\alias{theme_simple}
\alias{theme_donut}
\alias{theme_heatmap}
\title{Simple default themes for plots}
\usage{
theme_simple(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)

theme_donut(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)

theme_heatmap(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
Simple default themes for \code{ggplot2} plots generated by the
\code{pdallele} plotting functions. Largely based on \code{\link[ggplot2:ggtheme]{ggplot2::theme_gray()}} with
\code{\link[ggplot2:element]{ggplot2::element_text()}} replaced by \code{\link[ggtext:element_markdown]{ggtext::element_markdown()}} and other
elements modified or hidden as needed for donut plots and heat maps.
}
