% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleanup.R
\name{add_reference_gene_catalog}
\alias{add_reference_gene_catalog}
\title{Add Reference Gene Catalog information to .data by allele}
\usage{
add_reference_gene_catalog(.data, path = refgene_path)
}
\arguments{
\item{.data}{A dataframe or tibble to add to}

\item{path}{Path to an NCBI ReferenceGeneCatalog.txt file}
}
\description{
Filters after joining to remove invalid matches when a row contains
whitelisted_taxa. Coalesces empty 'gene' column with 'allele' column
to deal with missing Reference Gene genes
}
