% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cleanup.R
\name{clean_filter_alleles}
\alias{clean_filter_alleles}
\title{Clean and filter allele listings}
\usage{
clean_filter_alleles(data, filter, remove, deduplicate = TRUE)
}
\arguments{
\item{data}{A dataframe or tibble with an \code{allele} column}

\item{filter}{A character vector of \code{allele} names to keep. Uses grepl().
Partial matches are kept. Vector is collapsed with "|" (or operator).}

\item{remove}{A character vector of allele_call values to remove}

\item{deduplicate}{Should duplicate \code{allele}/\code{biosample} combinations be
removed?}
}
\value{
\code{data} with rows removed according to the selected parameters
}
\description{
The column \code{allele_call} contains codes representing to the
calling method used for a given allele. Some of these codes correspond to
"less than perfect" matches and may warrant removal depending on the scope
and goals of the analysis. Detailed explanations are available at:
\url{https://www.ncbi.nlm.nih.gov/pathogens/pathogens_help/#genotype-categories}

This function removes rows with \code{allele_call} matching a value in \code{remove}.

This function also (optionally) keeps only rows with \code{allele} matching a
value in \code{filter}, such as when only a specific subset of \code{allele} is
needed in the dataset.

Optionally, can deduplicate \code{allele}/\code{biosample} combinations. This can occur
when an isolate contains a large number of contigs resulting in a given
allele being listed several times for the isolate. Runs
\code{dplyr::distinct(biosample, allele, .keep_all = TRUE)}.
}
