% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{count_by_column}
\alias{count_by_column}
\title{Count by Columns}
\usage{
count_by_column(
  .data,
  ...,
  isolates = FALSE,
  sort = TRUE,
  name = "n",
  na_last = TRUE
)
}
\arguments{
\item{.data}{A data frame or tibble.}

\item{...}{\if{html}{\out{<data masking>}} columns to group by.}

\item{isolates}{Should isolates (BioSamples) be counted instead of alleles?}

\item{sort}{If TRUE, will show the largest groups at the top.}

\item{name}{The name of the new column in the output.}

\item{na_last}{Where to sort NA. If TRUE, NA is put last; if FALSE,
NA is put first; if NA, NA is dropped.}
}
\description{
Count by Columns
}
