% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{determine_combinations}
\alias{determine_combinations}
\title{Determine allele combinations}
\usage{
determine_combinations(data, filter_terms = NULL)
}
\arguments{
\item{data}{A data frame or tibble}

\item{filter_terms}{A character vector of (potentially partial) allele names
to keep. Collapsed with '|' and filtered with \code{\link[=grepl]{grepl()}}.}
}
\description{
Determine combinations of \code{allele} present in each isolate, concatenate a
list of alleles into a \code{combo} column, and remove isolates with only a single
allele.
}
