% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{donut_plot}
\alias{donut_plot}
\title{Generates a panel of donut plots of \code{fill} and \code{counts}}
\usage{
donut_plot(
  data,
  fill,
  counts,
  count_label = "Alleles",
  legend_title = "**Allele**",
  labels = TRUE,
  total = TRUE,
  palette = NULL
)
}
\arguments{
\item{data}{a data table or tibble}

\item{fill}{column containing levels/labels (e.g. alleles names) (rich text)}

\item{counts}{column containing counts for the levels}

\item{count_label}{label for counts (e.g. Alleles) (rich text)}

\item{legend_title}{label for the legend (rich text)}

\item{labels}{should callouts with count & percentage be shown for each arc?}

\item{total}{should total of \code{count} be displayed inside the donut?}

\item{palette}{specify a ggthemes::tableau_color_pal() color palette name}
}
\description{
See companion function donut_plot_panel() for a variation better suited for
comparison between groups of data
}
