% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{donut_plot_panel}
\alias{donut_plot_panel}
\title{Generates a panel of donut plots of \code{fill} and \code{counts} faceted by \code{facet}}
\usage{
donut_plot_panel(
  .data,
  fill = allele_markdown,
  counts = alleles,
  facet,
  count_label = "Alleles",
  legend_title = "**Allele**",
  total = TRUE,
  palette = NULL,
  nrow = NULL,
  ncol = NULL
)
}
\arguments{
\item{.data}{a data table or tibble}

\item{fill}{column containing levels/labels (e.g. alleles names) (rich text)}

\item{counts}{column containing counts for the levels}

\item{facet}{column to facet plots by (e.g. region)}

\item{count_label}{label for counts (e.g. Alleles) (rich text)}

\item{legend_title}{label for the legend (rich text)}

\item{total}{should total of \code{count} be displayed inside the donut?}

\item{palette}{specify a ggthemes::tableau_color_pal() color palette name}

\item{nrow}{number of facet rows}

\item{ncol}{number of facet columns}
}
\description{
See companion function donut_plot() for a variation better suited for
a single condition without faceting
}
