% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_identical_protein_groups}
\alias{download_identical_protein_groups}
\title{Download Identical Protein Groups Data using NCBI Entrez/Eutils}
\usage{
download_identical_protein_groups(accessions, n = 10, path)
}
\arguments{
\item{accessions}{a character vector of protein accession numbers to
determine Identical Protein Groups}

\item{n}{count of accession numbers per query}

\item{path}{directory to save downloaded files to}
}
\description{
A lower value for \code{n} will require more queries to NCBI servers but will reduce
redundant downloads if an IPG is associated with a large number of accession
numbers in the dataset.
}
