% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_metadata_clusters}
\alias{download_metadata_clusters}
\title{Download metadata and clusters for a specified PDG accession.version (FTP)}
\usage{
download_metadata_clusters(url, organism, version, path)
}
\arguments{
\item{url}{path to NCBI Pathogen Detection FTP "Results" directory}

\item{organism}{name of a valid organism/organism group}

\item{version}{a valid PDG accession.version}

\item{path}{directory to save downloaded files to}
}
\description{
Download metadata and clusters for a specified PDG accession.version (FTP)
}
