% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_microbigge_bq}
\alias{download_microbigge_bq}
\title{Download MicroBIGG-E Data from the Google Cloud Platform}
\usage{
download_microbigge_bq(
  billing = gcp_billing,
  path,
  taxgroup,
  sciname,
  element,
  debug = FALSE
)
}
\arguments{
\item{path}{directory to save downloaded files to}

\item{taxgroup}{an NCBI Pathogen Detection organism group to filter by}

\item{sciname}{scientific name of the organism of interest to filter by}

\item{element}{name of a gene/allele to filter by}

\item{debug}{should the query be previewed rather than executed?}

\item{gcp_billing}{GCP project identifier for billing/usage tracking}
}
\description{
See bigrquery help for more information about configuration and authentication
}
\details{
Note that taxgroup, sciname, and element all use SQL syntax for queries. Use
'\%' as a wildcard rather than '*'.
}
