% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ensure_directory}
\alias{ensure_directory}
\title{Ensure directory exists (and make if it doesn't)}
\usage{
ensure_directory(path)
}
\arguments{
\item{path}{directory path}
}
\description{
Ensure directory exists (and make if it doesn't)
}
