% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleanup.R
\name{filter_microbigge}
\alias{filter_microbigge}
\title{Filter imported MicroBigg-E data}
\usage{
filter_microbigge(data, coverage = 100L, identity = 90L, remove)
}
\arguments{
\item{data}{A dataframe or tibble}

\item{coverage}{Minimum percent coverage to keep}

\item{identity}{Minimum percent identity to keep}

\item{remove}{Character vector of \code{method} values to remove alleles}
}
\value{
\code{data} with rows removed according to the selected parameters
}
\description{
Removes allele calls that do not meet the minimum \code{coverage} and
\code{identity} thresholds. These values may vary based on the needs of a
particular analysis. Set both to \code{100} to allow only exact allele calls.

Specific allele calling methods can be excluded from the dataset by name in
\code{remove}. See <<>> for more information about allele calling methods used
in MicroBIGG-E data.
}
\details{
Isolates Browser (and thus amr.metadata.tsv files on the NCBI FTP
server) uses a 90\% \code{identity} threshold:
https://www.ncbi.nlm.nih.gov/pathogens/pathogens_help/#genotype-categories
}
