% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_organisms}
\alias{get_organisms}
\title{List Available Organisms (FTP)}
\usage{
get_organisms(url = ftp_results_base)
}
\arguments{
\item{url}{path to NCBI Pathogeen Detection FTP "Results" directory}
}
\description{
List Available Organisms (FTP)
}
