% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_organisms}
\alias{get_organisms}
\title{Download listing of available organisms}
\usage{
get_organisms(url = "ftp://ftp.ncbi.nlm.nih.gov/pathogen/Results/")
}
\arguments{
\item{url}{Path to the NCBI Pathogen Detection FTP Results directory}
}
\value{
A tibble with columns \code{organism} containing organism names and
\code{updated} containing the date of the most recent update.
}
\description{
Downloads listing of organisms currently available on the NCBI
Pathogen Detection FTP server.

The FTP directory can be accessed via browser at
\link{https://ftp.ncbi.nlm.nih.gov/pathogen/Results/}.
}
\seealso{
Other metadata and clusters: 
\code{\link{check_complete_pdg}()},
\code{\link{download_metadata_clusters}()},
\code{\link{get_versions}()}
}
\concept{metadata and clusters}
