% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_versions}
\alias{get_versions}
\title{Download listing of available PDG accession.version releases for a given organism}
\usage{
get_versions(url = "ftp://ftp.ncbi.nlm.nih.gov/pathogen/Results/", organism)
}
\arguments{
\item{url}{Path to the NCBI Pathogen Detection FTP Results directory}

\item{organism}{FTP directory name of an organism group included in the NCBI
Pathogen Detection Project}
}
\value{
A tibble with columns \code{version} containing available PDG
accession.version releases \code{updated} containing the date of the most recent
update.
}
\description{
Downloads listing of PDG (Pathogen Detection Organism Group)
accession.version releases currently available for \code{organism} on the NCBI
Pathogen Detection FTP server.
}
\seealso{
Other metadata and clusters: 
\code{\link{check_complete_pdg}()},
\code{\link{download_metadata_clusters}()},
\code{\link{get_organisms}()}
}
\concept{metadata and clusters}
