% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap_data}
\alias{heatmap_data}
\title{Prepare heat map data}
\usage{
heatmap_data(
  data,
  count,
  category,
  n = 5,
  facet,
  by_group = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{data}{A dataframe or tibble}

\item{count}{\if{html}{\out{<data-masking>}} Column to count (long axis)}

\item{category}{\if{html}{\out{<data-masking>}} Column to compare across (short axis)}

\item{n}{Number of most frequent \code{count} to include (i.e. top n =)}

\item{facet}{\if{html}{\out{<data-masking>}} Column to facet by}

\item{by_group}{Should \code{n} be aggregated by group or for the overall dataset?}

\item{na.rm}{Should \code{NA} values be removed from \code{count}, \code{category}, and \code{facet}?}
}
\value{
a tibble with ranks, proportions, and counts ready to plot using
\code{\link[=heatmap_plot]{heatmap_plot()}}
}
\description{
Prepare \code{data} for plotting as a heatmap with \code{\link[=heapmap_plot]{heapmap_plot()}}.

Count all instances of \code{count} by \code{category}, filter \code{n} most frequent \code{count}
either overall or \code{by_group}. Optionally prepare to facet by \code{facet} and
remove \code{NA} values.

Proportion refers to the proportion of isolates in a given \code{category} associated
with each \code{count}. Ranks are the \code{dense_rank(desc())} of the number of
isolates associated with each \code{count}.

The default setting \code{by_group = TRUE} will keep all \code{count} that appear
in the \code{n} most frequent in any \code{category}, typically leading to more than \code{n}
total values being kept. Conversely, \code{by_group = FALSE} will keep only the
\code{n} most frequent from the entire tibble, ensuring only \code{n} values in the
output.
}
\seealso{
\code{\link[=heatmap_plot]{heatmap_plot()}} to generate a \code{ggplot2} heatmap plot
}
