% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap_plot}
\alias{heatmap_plot}
\title{Prepare ggplot object of heatmap}
\usage{
heatmap_plot(
  data,
  count,
  group,
  facet,
  option = "viridis",
  wide = FALSE,
  rank = "**Rank**",
  prop = "**Proportion**",
  prop_breaks = c(1e-04, 5e-04, 0.001, 0.0025, 0.005, 0.01, 0.025, 0.05, 0.1, 0.2, 0.3,
    0.5, 0.75, 0.9),
  prop_labels = c("0.01\%", "0.05\%", "0.1\%", "0.25\%", "0.5\%", "1\%", "2.5\%", "5\%",
    "10\%", "20\%", "30\%", "50\%", "75\%", "90\%"),
  prop_barheight = 15,
  prop_limits = NULL,
  rank_breaks = c(1, 5, 10, 25, 50, 75, 100, 150, 200),
  rank_limits = NULL,
  rank_range = c(12, 1),
  text_size = 16
)
}
\arguments{
\item{data}{a dataframe or tibble}

\item{count}{\if{html}{\out{<data-masking>}} column to count (long axis)}

\item{group}{\if{html}{\out{<data-masking>}} column to compare across (short axis)}

\item{facet}{column to facet by.}

\item{option}{Color palette, passed to \code{viridis::scale_color_viridis()}.}

\item{wide}{should the output be wide or tall? See description.}

\item{rank}{title for the rank (size) legend}

\item{prop}{title for the proportion (color) legend}

\item{prop_breaks}{Proportion values to be displayed in the legend. Passed to
\code{ggplot2::continuous_scale(breaks = )}}

\item{prop_labels}{Labes corresponding to \code{prop_breaks}. Passed to
\code{ggplot2::continuous_scale(labels = )}}

\item{prop_barheight}{Height of the bar for the proportion legend Passed to
\code{ggplot2::guide_colorbar(barheight = )}}

\item{prop_limits}{Limits of the proportion color scale. Passed to
\code{ggplot2::continuous_scale(limits = )}}

\item{rank_breaks}{Rank values to be displayed in the legend. Passed to
\code{ggplot2::scale_size(breaks = )}.}

\item{rank_limits}{Limits of the rank size scale. Passed to
\code{ggplot2::scale_size(limits = )}}

\item{rank_range}{Minimum and maximum size of the circle. Passed to
\code{ggplot2::scale_size(range = )}}

\item{text_size}{Text size in plot. Passed to \code{ggplot2::element_text()}.}
}
\description{
Notes on Plot Orientation:
}
\details{
\code{wide = FALSE} places \code{count} on the Y-axis and \code{group} on the X-axis which
will result in "longer" plots when \code{count} has more levels than \code{group}.
Plots will facet vertically.

\code{wide = TRUE} swaps the axes and will result in "wider" plots when \code{count}
has more levels than \code{group}. Plots will facet horizontally.

Facet directions are fixed with plot orientation to avoid sections or to
avoid unexpected, unintended, and undesirable "tiling" of plots
}
