% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleanup.R
\name{import_cluster_list}
\alias{import_cluster_list}
\title{Import and add NCBI Pathogen Detection SNP Cluster data}
\usage{
import_cluster_list(data, path)
}
\arguments{
\item{data}{A dataframe or tibble containing a \code{biosample} column}

\item{path}{Path to an NCBI Pathogen Detection cluster_list.tsv file'}
}
\value{
\code{data} with column \code{pds} added
}
\description{
Import NCBI \strong{P}athogen \strong{D}etection \strong{S}NP Cluster (PDS) data from \code{path}
and add a \code{pds} column containing a PDS accession numbers to \code{data}.
}
