% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleanup.R
\name{import_ipg}
\alias{import_ipg}
\title{Import NCBI Identical Protein Groups data and add accession numbers and names
to MBE data}
\usage{
import_ipg(.data, path = ipg_path)
}
\arguments{
\item{.data}{a data table or tibble}

\item{path}{path to an NCBI IPG csv/tsv file}
}
\description{
This data can be obtained by downloading via the web interface, although it
seems rather inefficient to download the entire listing for all bla alleles
present in an organism.
}
