% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleanup.R
\name{import_isolates_browser_metadata}
\alias{import_isolates_browser_metadata}
\title{Import, Select, and Trim NCBI Pathogen Detection Isolates Browser Metadata}
\usage{
import_isolates_browser_metadata(path = isolates_path)
}
\arguments{
\item{path}{Path to an NCBI Isolates Browser
PDG_accession.version.metadata.tsv file}
}
\description{
Imports Isolated Browser metadataa, selects and renames desired columns, and
cleans up problematic quotation marks
}
