% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_process.R
\name{import_microbigge_gcp}
\alias{import_microbigge_gcp}
\title{Import and Select Columns from NCBI Pathogen Detection MicroBIGG-E file}
\usage{
import_microbigge_gcp(path)
}
\arguments{
\item{path}{Path to a MicroBIGG-E tsv file}
}
\value{
A \code{\link[=tibble]{tibble()}}. Values are described in Details.
}
\description{
Imports NCBI Pathogen Detection Microbial Browser for Identification of
Genetic and Genomic Elements (MicroBIGG-E) data and selects and renames
desired columns for downstream processing steps

Two variants are available:
\itemize{
\item \code{import_microbigge_ncbi()} imports files downloded from the NCBI Website
\item \code{import_microbigge_gcp()} imports files downloaded from Google Cloud Platform
}

Slightly different column names are used between NCBI and GCP, resulting in
two slightly different import functions.
}
\details{
The following columns are included in the output:\tabular{ll}{
   \strong{Source File} \tab \strong{Output Tibble} \cr
   scientific_name \tab species \cr
   protein_acc \tab protein \cr
   biosample_acc \tab biosample \cr
   collection_date \tab collection_date \cr
   asm_acc \tab assembly \cr
   target_acc \tab pdt \cr
   epi_type \tab sample_type \cr
   geo_loc_name \tab location \cr
   host \tab host \cr
   isolation_source \tab isolation_source \cr
   element_symbol \tab allele \cr
   element_name \tab name \cr
   class \tab class \cr
   subclass \tab subclass \cr
   amr_method \tab method \cr
   pct_ref_coverage \tab coverage \cr
   pct_ref_identity \tab identity \cr
   contig_acc \tab nucleotide \cr
   start_on_contig \tab nuc_start \cr
   end_on_contig \tab nuc_stop \cr
   strand \tab strand \cr
   ----------------- \tab ----------------- \cr
}
}
\section{Obtaining Data}{
 Up to 100,000 rows of data (potentially with an
additional 50 MB cap) can be downloaded via the web interface at:
\link{https://www.ncbi.nlm.nih.gov/pathogens/microbigge/}

Larger sets can be downloaded using Google Big Query. See instructions from
NCBI at \link{https://www.ncbi.nlm.nih.gov/pathogens/docs/microbigge_gcp/}
}

