% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleanup.R
\name{import_mlst}
\alias{import_mlst}
\title{Import and Add MLST Data}
\usage{
import_mlst(data, path)
}
\arguments{
\item{path}{Path to a (reprocessed) MLST CSV file}
}
\description{
Import and Add MLST Data
}
