% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleanup.R
\name{import_reference_gene_catalog}
\alias{import_reference_gene_catalog}
\title{Import Reference Gene Catalog to tibble}
\usage{
import_reference_gene_catalog(path)
}
\arguments{
\item{path}{Path to an NCBI ReferenceGeneCatalog.txt file}
}
\description{
Imports Reference Gene Catalog, renames columns, and removes unneeded data
}
