% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_process.R
\name{import_regions}
\alias{import_regions}
\title{Adds \code{region} column based on \code{location_broad} and a file}
\usage{
import_regions(data, path, warn = FALSE)
}
\arguments{
\item{data}{A dataframe or tibble}

\item{path}{Path to a csv file containing at least columns \code{country} and
\code{region}}

\item{warn}{Should a warning message with unmatched countries be shown?}
}
\value{
\code{data} with column \code{region} (and possibly others) added
}
\description{
Requires a csv file with a 'country' and 'region' column. Countries are
matched to location_broad using countrycode::countryname to account for names
given in as many languages as possible.
}
\details{
This function will also append any other columns in the file to the dataset
(using the column name from the file). This allows the inclusion of multiple
regions or including a region_markdown or similar column
}
