% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleanup.R
\name{na_if_tibble_chr}
\alias{na_if_tibble_chr}
\title{Replace character vector values with NA across all character columns}
\usage{
na_if_tibble_chr(.data, terms)
}
\arguments{
\item{.data}{A dataframe or tibble to modify}

\item{terms}{A character vector to replace values with NA_character_}
}
\description{
Replace character vector values with NA across all character columns
}
