% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleanup.R
\name{parse_bla_formatting}
\alias{parse_bla_formatting}
\title{Properly format  bla gene/allele names per ASM Journals guidelines}
\usage{
parse_bla_formatting(data)
}
\arguments{
\item{data}{A dataframe or tibble with an 'allele' and 'gene' column}
}
\description{
Add formatting for proper display of bla allele names in both markdown (used
by ggtext) and 'math' formatting (used by ggplot with expression()).
}
\details{
The format is "bla" in italic and gene/allele name in subscript, e.g. using
HTML tags: \if{html}{\out{<i>}}bla\if{html}{\out{</i>}}\if{html}{\out{<sub>}}PDC-3\if{html}{\out{</sub>}}. See PMID 35380458 "Consensus on
β-Lactamase Nomenclature" by Bush et al. for details
}
