% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_process.R
\name{parse_year}
\alias{parse_year}
\title{Parse \code{year} from \code{collection_date}}
\usage{
parse_year(data)
}
\arguments{
\item{data}{A dataframe or tibble with \code{collection_date} column}
}
\value{
\code{data} with column \code{year} added
}
\description{
Parse \code{year} from \code{collection_date}
}
