% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{pipe_inform}
\alias{pipe_inform}
\title{Pass data along the pipeline and rlang::inform()}
\usage{
pipe_inform(.data, ...)
}
\arguments{
\item{.data}{a dataframe or tibble}

\item{...}{parameters to pass to inform()}
}
\description{
Pass data along the pipeline and rlang::inform()
}
