% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{pipe_warn}
\alias{pipe_warn}
\title{Pass data along the pipeline and warn}
\usage{
pipe_warn(data, ...)
}
\arguments{
\item{data}{a dataframe or tibble}

\item{...}{parameters to pass to warn()}
}
\value{
unaltered \code{data}
}
\description{
Pass data along the pipeline and invoke rlang::warn() with the desired message
}
\seealso{
\code{\link[=pipe_inform]{pipe_inform()}}
}
