% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{remove_assigned_bla}
\alias{remove_assigned_bla}
\title{Remove assigned bla alleles}
\usage{
remove_assigned_bla(data)
}
\description{
Filter data to remove unassigned bla alleles
(i.e. those without an allele number)
}
\details{
#' @param data A data frame or tibble
}
