% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleanup.R
\name{reverse_geocode}
\alias{reverse_geocode}
\title{Reverse geocode "lat_lon" to countries and use to fill in missing "location"}
\usage{
reverse_geocode(data)
}
\arguments{
\item{data}{A dataframe or tibble to modify}
}
\description{
Reverse geocode "lat_lon" to countries and use to fill in missing "location"
}
