% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{select_ftp_organism}
\alias{select_ftp_organism}
\title{Interactively select an organism from the NCBI Pathogen Detection FTP directory}
\usage{
select_ftp_organism(url = ftp_results_base)
}
\arguments{
\item{url}{path to NCBI Pathogen Detection FTP "Results" directory}
}
\description{
Interactively select an organism from the NCBI Pathogen Detection FTP directory
}
