% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_process.R
\name{separate_genotypes}
\alias{separate_genotypes}
\title{Separate genotypes by row}
\usage{
separate_genotypes(data, include)
}
\arguments{
\item{data}{A data frame or tibble with at least one \code{amr}, \code{stress}, or \code{vir}
column}

\item{include}{Which types of alleles (\code{amr}, \code{stress}, \code{vir}) should be
kept? Separate values with '|'.}
}
\value{
\code{data} with columns \code{allele}, \code{allele_type}, and \code{method} added
}
\description{
Uses separate_rows to separate genotype columns (\code{amr},
\code{stress}, \code{vir}) into rows, with each row corresponding to a single allele.
Three columns of allele details are added:
\itemize{
\item \code{allele} corresponding to the allele name
\item \code{allele_type} corresponding to the source column (\code{amr}, \code{stress}, or \code{vir})
\item \code{method} corresponding to codes used by the NCBI Pathogen Detection Project to describe the method used for calling a given allele. See \code{\link[=clean_filter_alleles]{clean_filter_alleles()}} for details.
}
}
