% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_process.R
\name{split_location}
\alias{split_location}
\title{Split location into location_broad and location_detail}
\usage{
split_location(data)
}
\arguments{
\item{data}{A data frame or tibble}
}
\value{
\code{data} with columns \code{location_broad} and \code{location_detail}
added
}
\description{
Splits INSDC/BioSample geo_loc_name style country:detailed_location format
into location_broad (country) and location_detail (city/area).
}
