% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_process.R
\name{split_location}
\alias{split_location}
\title{Split location into location_broad and location_detail}
\usage{
split_location(data, custom_match = NULL)
}
\arguments{
\item{data}{A data frame or tibble}

\item{custom_match}{A named character vector correlating original country
names to standardized country names. Passed to \code{custom_match} of
\code{\link[=standardize_countries]{standardize_countries()}}.}
}
\value{
\code{data} with columns \code{location_broad} and \code{location_detail}
added
}
\description{
Splits INSDC/BioSample geo_loc_name style country:detailed_location format
into \code{location_broad} (country) and \code{location_detail} (city/area) and adds
\code{standardized_country} which attempts to account for spelling and language
differences.
}
