% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{top_n_overall}
\alias{top_n_overall}
\alias{top_n_across_groups}
\alias{top_n_by_group}
\title{Rank, sort, and combine values}
\usage{
top_n_overall(
  data,
  ...,
  count,
  n = 10,
  desc = TRUE,
  other = "Other",
  relevel = "name",
  other_pos = "last"
)

top_n_across_groups(
  data,
  ...,
  count,
  n = 10,
  desc = TRUE,
  other = "Other",
  relevel = "name",
  other_pos = "last"
)

top_n_by_group(
  data,
  ...,
  count,
  n = 10,
  desc = TRUE,
  other = "Other",
  relevel = "name",
  other_pos = "last"
)
}
\arguments{
\item{data}{a dataframe or tibble}

\item{...}{\if{html}{\out{<data-masking>}} columns of labels to be ranked}

\item{count}{column of values (typically counts) to rank}

\item{n}{number of rows to keep per group (all other values will be collapsed
into an \code{other} row, meaning n+1 total rows will appear in out unless
\code{other_pos} is set to "drop"}

\item{desc}{should the n highest values of \code{count} be kept?}

\item{other}{character label for the "other" group}

\item{relevel}{should factors in \code{...} be releveled in alphanumeric order?}

\item{other_pos}{where should "other" sort?}
}
\description{
These functions are primarailly designed to prepare data for use
in \code{\link[=donut_plot]{donut_plot()}} and serve three main purposes:
\enumerate{
\item Determine the \code{n} largest (or smallest) counts
\item Combine remaining counts into \code{other}
\item Optionally, relevel factor levels using [\code{\link[=relevel_numeric]{relevel_numeric()}}]
}

They differ in the handling of groups:
\itemize{
\item \code{top_n_overall()} determines the "top n" of the entire dataset and keeps those labels across all groups. Up to \code{n} labels + \code{other} will be kept.
\item \code{top_n_by_group()} determines the "top n" on a group-by-group basis and keeps the "top n" labels for each group independently. The total number of labels kept will vary, with a maximum of \code{n} \* length(groups) + \code{other}
\item \code{top_n_across_groups()} determines the "top n" on a group-by-group basis and keeps labels appearing in the "top n" of any group across all groups. Up to \code{n} labels + \code{other} will be kept.
}

All values that aren't 'kept' are combined into \code{other}.

No tie-breaking is performed. In the event of a tie for the nth row, the
first row by arrange() is kept.
}
