% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{top_n_overall}
\alias{top_n_overall}
\title{Keep the \code{n} highest (or lowest) \code{count} rows overall for each group and
combine the rest into \code{other}}
\usage{
top_n_overall(
  data,
  ...,
  count,
  n = 10,
  desc = TRUE,
  other = "Other",
  relevel = "name",
  other_pos = "last"
)
}
\arguments{
\item{data}{a dataframe or tibble}

\item{...}{\if{html}{\out{<data-masking>}} columns of labels to be ranked}

\item{count}{column of values (typically counts) to rank}

\item{n}{number of rows to keep per group (all other values will be collapsed
into an \code{other} row, meaning n+1 total rows will appear in out unless
\code{other_pos} is set to "drop"}

\item{desc}{should the n highest values of \code{count} be kept?}

\item{other}{character label for the "other" group}

\item{relevel}{should factors in \code{...} be releveled in alphanumeric order?}

\item{other_pos}{where should "other" sort?}
}
\description{
This function keeps the "top n" overall for each group. See the companion function
top_n_across_groups() to keep the "top n" in each group across all groups.
}
\details{
Note that no tie-breaking is performed. In the event of a tie for the nth
row, the first row by arrange() is kept.
}
